Convert to Jekyll website
The goal of the static website was to get it up and running. Jekyll was
not used because there were too many choices in templates, etc. Now
that posts are coming, the static web is manual, and producing too many
errors (dates, titles, etc.) because of the copying.

The goal now is to get a sitemap and rss feed generated so the blog can
be submitted to the search engines. After finding a half dozen typos in
content (not related to the text) I figured now was the time to switch.

The same post index method is used (via permalinks) because I really
like it. Also, the work is similar because in the static website this
was done for all href in the anchors tags.