Add RSS feed
Part of the reasoning for migrating to Jekyll was to get the RSS feed
active without having to manually create it. The attempts at that were
filled with errors (and I discovered more errors inside the website as
well). Manual got it started but was never a good idea (or long term).