Switch to Jekyll syntax highlighting
The switch to Jekyll based syntax highlighting removes the client
side JavaScript in favor of Jekyll pre-processing. The `github.css`
was included to give the code a familiar look and feel.

The posts that contained code needed to have their extensions changed
to `.md` for markdown processing. This is where the syntax highlighting
is done. The CSS renders the generated styles (like a theme).