Add Google Analytics
Removed the plausible analytics tag in favor of Google due to cost. The blog
will take 3-6 months before any traffic *should* start to appear and I do
not want to pay anything for the waiting period. I may chanage the provider
at a later time but for now, let's keep it cheap.