Runbooks Document the Failures in Your System
Time has come this weekend to submit another ticket for a certain system,
that our software depends upon, that will be down due to patching. This
post is cathartic but also a realization that it will never be fixed. All
because of yet another great Stay Sassy post on runbooks.