Arrange/Act/Assert post
This is part#2 on unit testing. It goes over the aaa pattern and how
to write the test. The arrange section is the largest as it covers
test doubles (I know, that could be a post on its own).

This was getting long, and rewritten 3 times, but I think I finally
hit the appropriate content.