Generate the TAOCP from tags
I started to see me making tons of errors in keeping the posts up to
date so I used tags instead. A quick search turned up the `tags` code
on the `taocp.html` layout page. It may not be efficient but it works
and this is Jekyll so publishing time is not exactly instantaneous to
begin with.